+++ tests/testthat/test-estimate_contrasts.R
@@ -364,6 +364,51 @@ test_that("estimate_contrasts - p.adjust", {
     p_adjust = "scheffe",
     backend = "emmeans"
   ))
+
+  skip_if_not_installed("mvtnorm")
+  dat <- iris
+  dat$fac <- ifelse(dat$Sepal.Length < 5.8, "A", "B")
+  model <- lm(Sepal.Width ~ Species * fac, data = dat)
+  set.seed(123)
+  out <- estimate_contrasts(model, c("Species", "fac"), p_adjust = "sup-t")
+  expect_equal(
+    out$p,
+    c(
+      0.44686, 0, 0, 0.00048, 0, 0.00086, 0.0086, 0.00328, 0.02718,
+      0.10619, 0.99998, 0.00037, 0.71917, 0.51122, 0.28497
+    ),
+    tolerance = 1e-3
+  )
+  expect_equal(
+    out$CI_low,
+    c(
+      -0.3461, -1.02306, -0.76301, -1.36376, -0.60638, -2.30875,
+      -2.06719, -2.46281, -1.93229, -0.0281, -0.60143, 0.12295, -0.82718,
+      -0.08844, -0.15023
+    ),
+    tolerance = 1e-3
+  )
+
+  skip_if_not_installed("glmmTMB")
+  d <- glmmTMB::Salamanders
+  model <- suppressWarnings(glmmTMB::glmmTMB(
+    count ~ mined + spp + (1 | site),
+    ziformula = ~mined,
+    family = poisson,
+    data = d
+  ))
+  set.seed(123)
+  out <- head(estimate_contrasts(model, "spp", by = "mined", p_adjust = "sup-t"))
+  expect_equal(
+    out$p,
+    c(0.00259, 0.59628, 0.18012, 0.00475, 0.00674, 0.99467),
+    tolerance = 1e-3
+  )
+  expect_equal(
+    out$CI_low,
+    c(-0.29028, -0.04581, -0.21504, 0.04106, 0.03308, -0.0816),
+    tolerance = 1e-3
+  )
 })